// src/evaluation/dtos/evaludation.controller.dto.ts
export class createEvaludationResponseDTO {
  nameOfLecture: string;
  nameOfProfessor: string;
  totalRate: number;
  assignmentRate: number;
  markRate: number;
}

export class getEvaludationResponseDTO {
  nameOfLecture: string;
  nameOfProfessor: string;
  totalRate: number;
  assignmentRate: number;
  markRate: number;
}